initial consolidation
