+++ packages/browser/src/index.ts
@@ -2,7 +2,6 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-export * from "./CallbackHandler";
 export * from "./Client";
 export * from "./LoggerCategory";
