+++ packages/browser/src/test/BrowserAuthorizationClient.test.ts
@@ -1,18 +1,15 @@
 /*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 
 import { assert } from "chai";
 import { BrowserAuthorizationClient } from "../Client";
 import { getImsAuthority } from "../utils";
-
-import type { BrowserAuthorizationClientConfiguration } from "../Client";
+import { BrowserAuthorizationClientConfiguration } from "../types";
 
 describe("BrowserAuthorizationClient", () => {
-
   describe("#constructor", () => {
-
     const TEST_AUTHORITY = "https://test.authority.com";
 
     let testClient: BrowserAuthorizationClient;
@@ -28,6 +25,7 @@ describe("BrowserAuthorizationClient", () => {
         responseType: "test_responseType",
         noSilentSignInOnAppStartup: false,
         silentRedirectUri: "test_silentRedirectUri",
+        responseMode: "query",
       };
 
       testConfig = {
@@ -44,7 +42,10 @@ describe("BrowserAuthorizationClient", () => {
       assert.equal(settings.authority, TEST_AUTHORITY);
       assert.equal(settings.clientId, "test_clientId");
       assert.equal(settings.redirectUri, "test_redirectUri");
-      assert.equal(settings.postSignoutRedirectUri, "test_postSignoutRedirectUri");
+      assert.equal(
+        settings.postSignoutRedirectUri,
+        "test_postSignoutRedirectUri"
+      );
       assert.equal(settings.scope, "test_scope");
       assert.equal(settings.responseType, "test_responseType");
       assert.equal(settings.noSilentSignInOnAppStartup, false);
@@ -80,6 +81,64 @@ describe("BrowserAuthorizationClient", () => {
 
       // getImsAuthority manages the value of the default authority
       assert.equal(client.authorityUrl, getImsAuthority());
+    });
+
+    it("_basicSettings contains passed in configuration", () => {
+      const settings = testClient["_basicSettings"]; // eslint-disable-line @typescript-eslint/dot-notation
+
+      assert.equal(settings.authority, TEST_AUTHORITY);
+      assert.equal(settings.clientId, "test_clientId");
+      assert.equal(settings.redirectUri, "test_redirectUri");
+      assert.equal(settings.responseMode, "query");
+    });
+
+    it("given authority is used", () => {
+      assert.equal(testClient.authorityUrl, TEST_AUTHORITY);
+    });
+
+    it("given authority is used and environment prefix is ignored", async () => {
+      process.env.IMJS_URL_PREFIX = "prefix-";
+
+      assert.equal(testClient.authorityUrl, TEST_AUTHORITY);
+    });
+
+    it("given authority is when no prefix is defined", async () => {
+      process.env.IMJS_URL_PREFIX = "";
+
+      assert.equal(testClient.authorityUrl, TEST_AUTHORITY);
+    });
+
+    it("default authority is used when none is given", () => {
+      const client = new BrowserAuthorizationClient(testConfigWithoutAuthority);
+
+      // getImsAuthority manages the value of the default authority
+      assert.equal(client.authorityUrl, getImsAuthority());
+    });
+
+    it("default authority is used and when none is given and uses environment prefix", () => {
+      process.env.IMJS_URL_PREFIX = "prefix-";
+      const client = new BrowserAuthorizationClient(testConfigWithoutAuthority);
+
+      // getImsAuthority manages the value of the default authority
+      assert.equal(client.authorityUrl, getImsAuthority());
+    });
+
+    it('successfully sets "query" as response mode', () => {
+      const client = new BrowserAuthorizationClient({
+        ...testConfig,
+        responseMode: "query",
+      });
+
+      assert.equal(client["_basicSettings"].responseMode, "query"); // eslint-disable-line @typescript-eslint/dot-notation
+    });
+
+    it('successfully sets "fragment" as response mode', () => {
+      const client = new BrowserAuthorizationClient({
+        ...testConfig,
+        responseMode: "fragment",
+      });
+
+      assert.equal(client["_basicSettings"].responseMode, "fragment"); // eslint-disable-line @typescript-eslint/dot-notation
     });
   });
 });
